python-backend/optimiser_service/requirements.txt
requests
PyPortfolioOpt
scikit-learn
flask
python-dotenv
pymongo